refactor(domain): use dataclass (#75)
refactor(repo): split models
