+++ heliotrope/database/base.py
@@ -1,0 +1,3 @@
+from sqlalchemy.orm.decl_api import registry
+
+mapper_registry = registry()
