--- heliotrope/shared/domain_model.py
@@ -1,13 +1,0 @@
-from abc import ABCMeta, abstractmethod
-from typing import Any
-
-
-class DomainModel(metaclass=ABCMeta):
-    @classmethod
-    @abstractmethod
-    def from_dict(cls, d: Any) -> Any:
-        raise NotImplementedError
-
-    @abstractmethod
-    def to_dict(self) -> Any:
-        raise NotImplementedError
